feat: overhaul sip_trunking validation and fix issues with values not being removed on update
Improve validation of the resources and data sources
- validate SID format
- add validation for trunk friendly name and transfer mode
- add validation for origination url priority, sip url and weight

Add default values
- trunk
-- cnam_lookup_enabled, recording mode, recording trim, secure and transfer_mode

These default values were included to allow me to address an issue when the provider was no clearing down an argument value in Twilio when it was removed from the Terraform configuration. This is a change in the behaviour of the provider **breaking change**. The following arguments will be set to the following:
- trunk
cnam_lookup_enabled -> false
disaster_recovery_method -> ""
recording mode -> do-not-record (when block is still present)
recording trim -> do-not-trim (when block is still present)
secure -> false
transfer_mode -> disable-all

The deprecated `sid` argument on the `twilio_sip_trunking_phone_number` resources has now been removed and the `phone_number_sid` argument has now become mandatory.