feat: overhaul account validation and fix issues with values not being removed on update
If the `friendly_name` is managed via Terraform and the `friendly_name` is removed from the configuration file. The old value will be retained on the next apply.

Improve validation of the resources and data sources
- validate SID format
- add validation for sub account status
- add validation for address customer name, street, city, region, postal code & iso country

Add default values
- address
-- emergency_enabled

These default values were included to allow me to address an issue when the provider was no clearing down an argument value in Twilio when it was removed from the Terraform configuration. This is a change in the behaviour of the provider **breaking change**. The following arguments will be set to the following:
- address
-- friendly_name -> ""
-- street_secondary -> ""
-- emergency_enabled -> false
- sub account
-- status -> "active"