feat: overhaul serverless validation and fix issues with values not being removed on update
If the `dependencies` or `runtime` are managed via Terraform and the values are removed from the configuration file. The old values will be retained on the next apply.

Improve validation of the resources and data sources
- validate SID format
- add validation for variable key and value
- add validation for service friendly name and unique name
- add validation for function friendly name, content type and path
- add validation for environment unique name and domain suffix
- add validation for asset friendly name, content type and path

Add default values
- service
-- include_credentials
-- ui_editable

These default values were included to allow me to address an issue when the provider was no clearing down an argument value in Twilio when it was removed from the Terraform configuration. This is a change in the behaviour of the provider **breaking change**. The following arguments will be set to the following:
- service
-- ui_editable -> false
-- include_credentials -> true
- environment
-- domain_suffix -> ""