chore: revert behaviour of replacing sid with service defaults (where appropiate)
This change would have been a fundamental change and would have led to inconistency in the provider.
Instead the approach of making the user explicity set the new SID is used instead.

Few documentation improvements
Revert optional helpers back to previous implementation and add explicit functions when empty strings or 0 is used