feat: overhaul autopilot validation and apply sensible defaults
Improve validation of the resources and data sources. Added default values. These default values were included to allow me to address an issue when the provider was no clearing down an argument value in Twilio when it was removed from the Terraform configuration. This is a change in the behaviour of the provider **breaking change**.