fix: issue when the Twilio API was returning function and asset versions in a random order
Previously when creating a serverless build with multiple functions or asset versions that were created simultaneously, the Twilio API could return these in any order. As the order may not match the order the asset or function versions were supplied, Terraform will detect this as a drift and the build is re-created on the next deployment. A diff suppression has been added to the `sid` argument in the `asset_version` and `function_version` block to check if the SID was present in the list, just in a different position

fixes: #30